Fix period calculations
